Extract HandshakeUtils for code reuse
