+++ modules/client-common/src/main/java/org/apache/ignite/internal/client/proto/HandshakeUtils.java
@@ -1,0 +1,65 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.ignite.internal.client.proto;
+
+import java.util.BitSet;
+import java.util.EnumMap;
+import java.util.Map;
+
+/**
+ * Utils for client protocol handshake handling.
+ */
+public class HandshakeUtils {
+    /**
+     * Unpacks extensions.
+     *
+     * @param unpacker Unpacker.
+     * @return Extensions.
+     */
+    public static Map<HandshakeExtension, Object> unpackExtensions(ClientMessageUnpacker unpacker) {
+        EnumMap<HandshakeExtension, Object> extensions = new EnumMap<>(HandshakeExtension.class);
+        int mapSize = unpacker.unpackInt();
+        for (int i = 0; i < mapSize; i++) {
+            HandshakeExtension handshakeExtension = HandshakeExtension.fromKey(unpacker.unpackString());
+            if (handshakeExtension != null) {
+                extensions.put(handshakeExtension, unpackExtensionValue(handshakeExtension, unpacker));
+            }
+        }
+        return extensions;
+    }
+
+    /**
+     * Unpacks features.
+     *
+     * @param unpacker Unpacker.
+     * @return Features.
+     */
+    public static BitSet unpackFeatures(ClientMessageUnpacker unpacker) {
+        var featuresLen = unpacker.unpackBinaryHeader();
+        var features = BitSet.valueOf(unpacker.readPayload(featuresLen));
+    }
+
+    private static Object unpackExtensionValue(HandshakeExtension handshakeExtension, ClientMessageUnpacker unpacker) {
+        Class<?> type = handshakeExtension.valueType();
+        if (type == String.class) {
+            return unpacker.unpackString();
+        } else {
+            throw new IllegalArgumentException("Unsupported extension type: " + type.getName());
+        }
+    }
+}
