+++ modules/client-handler/src/main/java/org/apache/ignite/client/handler/ClientInboundMessageHandler.java
@@ -31,7 +31,6 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.ChannelInboundHandlerAdapter;
 import io.netty.handler.codec.DecoderException;
 import java.util.BitSet;
-import java.util.EnumMap;
 import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
@@ -104,6 +103,7 @@ import org.apache.ignite.internal.client.proto.ClientMessageUnpacker;
 import org.apache.ignite.internal.client.proto.ClientOp;
 import org.apache.ignite.internal.client.proto.ErrorExtensions;
 import org.apache.ignite.internal.client.proto.HandshakeExtension;
+import org.apache.ignite.internal.client.proto.HandshakeUtils;
 import org.apache.ignite.internal.client.proto.ProtocolVersion;
 import org.apache.ignite.internal.client.proto.ResponseFlags;
 import org.apache.ignite.internal.compute.IgniteComputeInternal;
@@ -364,11 +364,10 @@ public class ClientInboundMessageHandler extends ChannelInboundHandlerAdapter im
                         + clientVer.major() + "." + clientVer.minor() + "." + clientVer.patch());
             }
 
-            var clientCode = unpacker.unpackInt();
-            var featuresLen = unpacker.unpackBinaryHeader();
-            var features = BitSet.valueOf(unpacker.readPayload(featuresLen));
+            int clientCode = unpacker.unpackInt();
 
-            Map<HandshakeExtension, Object> extensions = extractExtensions(unpacker);
+            BitSet features = HandshakeUtils.unpackFeatures(unpacker);
+            Map<HandshakeExtension, Object> extensions = HandshakeUtils.unpackExtensions(unpacker);
 
             authenticationManager
                     .authenticateAsync(createAuthenticationRequest(extensions))
@@ -969,27 +968,6 @@ public class ClientInboundMessageHandler extends ChannelInboundHandlerAdapter im
                 + ctx.channel().remoteAddress() + "]: " + cause.getMessage(), cause);
 
         ctx.close();
-    }
-
-    private static Map<HandshakeExtension, Object> extractExtensions(ClientMessageUnpacker unpacker) {
-        EnumMap<HandshakeExtension, Object> extensions = new EnumMap<>(HandshakeExtension.class);
-        int mapSize = unpacker.unpackInt();
-        for (int i = 0; i < mapSize; i++) {
-            HandshakeExtension handshakeExtension = HandshakeExtension.fromKey(unpacker.unpackString());
-            if (handshakeExtension != null) {
-                extensions.put(handshakeExtension, unpackExtensionValue(handshakeExtension, unpacker));
-            }
-        }
-        return extensions;
-    }
-
-    private static Object unpackExtensionValue(HandshakeExtension handshakeExtension, ClientMessageUnpacker unpacker) {
-        Class<?> type = handshakeExtension.valueType();
-        if (type == String.class) {
-            return unpacker.unpackString();
-        } else {
-            throw new IllegalArgumentException("Unsupported extension type: " + type.getName());
-        }
     }
 
     private static <T> @Nullable T findException(Throwable e, Class<T> cls) {
